server/db/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
httpx
icalendar
requests
pymongo
python-telegram-bot


